backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
azure-monitor-query==1.4.0
azure-identity==1.19.0
pydantic==2.9.2
pydantic-settings==2.6.1
cachetools==5.5.0
python-dateutil==2.9.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
